requirements.txt
sanic==23.6.0
sanic-routing==23.6.0
sanic-jwt==1.8.0
tortoise-orm==0.20.0
asyncpg==0.28.0
Jinja2==3.1.2
ipaddress==1.0.23
geoip2==4.7.0
environs==9.5.0
httpx==0.24.1